requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
pillow
cloudinary
